+++ moban/jinja2/filters/text.py
@@ -1,18 +1,19 @@
 import re
-from moban.extensions import JinjaFilter
+
+from moban.jinja2.extensions import JinjaFilter
 
 
 @JinjaFilter()
 def split_length(input_line, length):
     start = 0
     limit = length
-    line = re.sub("\s+", " ", input_line)
+    line = re.sub(r"\s+", " ", input_line)
     line_length = len(line)
     if line_length <= length:
         yield line
     else:
         while True:
-            if " " in line[start : start + limit]: # flake8: noqa 
+            if " " in line[start : start + limit]:  # noqa
                 # go back and find a space
                 while limit > 0 and line[start + limit] != " ":
                     limit -= 1
@@ -24,7 +25,7 @@ def split_length(input_line, length):
                 ] != " ":
                     limit += 1
 
-            yield line[start : start + limit]  # flake8: noqa
+            yield line[start : start + limit]  # noqa
             start = start + limit + 1
             limit = length
             if len(line[start:]) < length or start + limit >= len(line):