+++ tests/integration_tests/test_command_line_options.py
@@ -1,9 +1,9 @@
 import os
 import sys
 from shutil import copyfile
-from moban.main import main
+
 from mock import patch
-from nose.tools import raises, assert_raises, eq_
+from nose.tools import eq_, raises, assert_raises
 
 
 class TestCustomOptions:
@@ -11,12 +11,12 @@ class TestCustomOptions:
         self.config_file = "config.yaml"
         with open(self.config_file, "w") as f:
             f.write("hello: world")
-        self.patcher = patch(
-            "moban.engine.verify_the_existence_of_directories"
+        self.patcher1 = patch(
+            "moban.plugins.verify_the_existence_of_directories"
         )
-        self.patcher.start()
+        self.patcher1.start()
 
-    @patch("moban.engine.Engine.render_to_file")
+    @patch("moban.plugins.BaseEngine.render_to_file")
     def test_custom_options(self, fake_template_doer):
         test_args = [
             "moban",
@@ -30,15 +30,19 @@ class TestCustomOptions:
             "a.jj2",
         ]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             fake_template_doer.assert_called_with(
                 "a.jj2", "config.yaml", "moban.output"
             )
 
-    @patch("moban.engine.Engine.render_to_file")
+    @patch("moban.plugins.BaseEngine.render_to_file")
     def test_minimal_options(self, fake_template_doer):
         test_args = ["moban", "-c", self.config_file, "-t", "a.jj2"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             fake_template_doer.assert_called_with(
                 "a.jj2", "config.yaml", "moban.output"
@@ -48,10 +52,12 @@ class TestCustomOptions:
     def test_missing_template(self):
         test_args = ["moban", "-c", self.config_file]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
 
     def tearDown(self):
-        self.patcher.stop()
+        self.patcher1.stop()
         os.unlink(self.config_file)
 
 
@@ -60,15 +66,17 @@ class TestOptions:
         self.config_file = "data.yml"
         with open(self.config_file, "w") as f:
             f.write("hello: world")
-        self.patcher = patch(
-            "moban.engine.verify_the_existence_of_directories"
+        self.patcher1 = patch(
+            "moban.plugins.verify_the_existence_of_directories"
         )
-        self.patcher.start()
+        self.patcher1.start()
 
-    @patch("moban.engine.Engine.render_to_file")
+    @patch("moban.plugins.BaseEngine.render_to_file")
     def test_default_options(self, fake_template_doer):
         test_args = ["moban", "-t", "a.jj2"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             fake_template_doer.assert_called_with(
                 "a.jj2", "data.yml", "moban.output"
@@ -78,6 +86,8 @@ class TestOptions:
     def test_no_argments(self):
         test_args = ["moban"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
 
     @raises(SystemExit)
@@ -87,7 +97,7 @@ class TestOptions:
             main()
 
     def tearDown(self):
-        self.patcher.stop()
+        self.patcher1.stop()
         os.unlink(self.config_file)
 
 
@@ -95,6 +105,8 @@ class TestOptions:
 def test_missing_configuration():
     test_args = ["moban", "-t", "a.jj2"]
     with patch.object(sys, "argv", test_args):
+        from moban.main import main
+
         main()
 
 
@@ -108,15 +120,17 @@ class TestNoOptions:
         self.data_file = "data.yaml"
         with open(self.data_file, "w") as f:
             f.write("hello: world")
-        self.patcher = patch(
-            "moban.engine.verify_the_existence_of_directories"
+        self.patcher1 = patch(
+            "moban.plugins.verify_the_existence_of_directories"
         )
-        self.patcher.start()
+        self.patcher1.start()
 
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.plugins.BaseEngine.render_to_files")
     def test_single_command(self, fake_template_doer):
         test_args = ["moban"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             call_args = list(fake_template_doer.call_args[0][0])
             eq_(
@@ -127,10 +141,12 @@ class TestNoOptions:
                 ],
             )
 
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.plugins.BaseEngine.render_to_files")
     def test_single_command_with_a_few_options(self, fake_template_doer):
         test_args = ["moban", "-t", "abc.jj2", "-o", "xyz.output"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             call_args = list(fake_template_doer.call_args[0][0])
             eq_(
@@ -142,7 +158,7 @@ class TestNoOptions:
                 ],
             )
 
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.plugins.BaseEngine.render_to_files")
     def test_single_command_with_options(self, fake_template_doer):
         test_args = [
             "moban",
@@ -154,6 +170,8 @@ class TestNoOptions:
             "xyz.output",
         ]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             call_args = list(fake_template_doer.call_args[0][0])
             eq_(
@@ -166,16 +184,18 @@ class TestNoOptions:
             )
 
     @raises(Exception)
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.plugins.BaseEngine.render_to_files")
     def test_single_command_without_output_option(self, fake_template_doer):
         test_args = ["moban", "-t", "abc.jj2"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
 
     def tearDown(self):
         os.unlink(self.config_file)
         os.unlink(self.data_file)
-        self.patcher.stop()
+        self.patcher1.stop()
 
 
 class TestNoOptions2:
@@ -188,15 +208,17 @@ class TestNoOptions2:
         self.data_file = "data.yaml"
         with open(self.data_file, "w") as f:
             f.write("hello: world")
-        self.patcher = patch(
-            "moban.engine.verify_the_existence_of_directories"
+        self.patcher1 = patch(
+            "moban.plugins.verify_the_existence_of_directories"
         )
-        self.patcher.start()
+        self.patcher1.start()
 
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.plugins.BaseEngine.render_to_files")
     def test_single_command(self, fake_template_doer):
         test_args = ["moban"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             call_args = list(fake_template_doer.call_args[0][0])
             eq_(
@@ -208,7 +230,7 @@ class TestNoOptions2:
             )
 
     def tearDown(self):
-        self.patcher.stop()
+        self.patcher1.stop()
         os.unlink(self.config_file)
         os.unlink(self.data_file)
 
@@ -222,15 +244,17 @@ class TestCustomMobanFile:
         self.data_file = "data.yaml"
         with open(self.data_file, "w") as f:
             f.write("hello: world")
-        self.patcher = patch(
-            "moban.engine.verify_the_existence_of_directories"
+        self.patcher1 = patch(
+            "moban.plugins.verify_the_existence_of_directories"
         )
-        self.patcher.start()
+        self.patcher1.start()
 
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.plugins.BaseEngine.render_to_files")
     def test_single_command(self, fake_template_doer):
         test_args = ["moban", "-m", self.config_file]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             call_args = list(fake_template_doer.call_args[0][0])
             eq_(
@@ -242,17 +266,19 @@ class TestCustomMobanFile:
             )
 
     def tearDown(self):
-        self.patcher.stop()
+        self.patcher1.stop()
         os.unlink(self.config_file)
         os.unlink(self.data_file)
 
 
-@patch("moban.engine.verify_the_existence_of_directories")
+@patch("moban.plugins.verify_the_existence_of_directories")
 def test_duplicated_targets_in_moban_file(fake_verify):
     config_file = "duplicated.moban.yml"
     copyfile(os.path.join("tests", "fixtures", config_file), ".moban.yml")
     test_args = ["moban"]
     with patch.object(sys, "argv", test_args):
+        from moban.main import main
+
         assert_raises(SystemExit, main)
     os.unlink(".moban.yml")
 
@@ -262,30 +288,36 @@ class TestInvalidMobanFile:
         self.config_file = ".moban.yml"
 
     @raises(SystemExit)
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.plugins.BaseEngine.render_to_files")
     def test_no_configuration(self, fake_template_doer):
         with open(self.config_file, "w") as f:
             f.write("")
         test_args = ["moban"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
 
     @raises(SystemExit)
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.plugins.BaseEngine.render_to_files")
     def test_no_configuration_2(self, fake_template_doer):
         with open(self.config_file, "w") as f:
             f.write("not: related")
         test_args = ["moban"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
 
     @raises(SystemExit)
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.plugins.BaseEngine.render_to_files")
     def test_no_targets(self, fake_template_doer):
         with open(self.config_file, "w") as f:
             f.write("configuration: test")
         test_args = ["moban"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
 
     def tearDown(self):
@@ -301,29 +333,31 @@ class TestComplexOptions:
         self.data_file = "data.yaml"
         with open(self.data_file, "w") as f:
             f.write("hello: world")
-        self.patcher = patch(
-            "moban.engine.verify_the_existence_of_directories"
+        self.patcher1 = patch(
+            "moban.plugins.verify_the_existence_of_directories"
         )
-        self.patcher.start()
+        self.patcher1.start()
 
-    @patch("moban.engine.Engine.render_to_files")
-    def test_single_command(self, fake_template_doer):
+    def test_single_command(self):
         test_args = ["moban"]
         with patch.object(sys, "argv", test_args):
-            main()
-            call_args = list(fake_template_doer.call_args[0][0])
-            eq_(
-                call_args,
-                [
-                    ("README.rst.jj2", "custom-data.yaml", "README.rst"),
-                    ("setup.py.jj2", "data.yml", "setup.py"),
-                ],
-            )
+            from moban.main import main
+
+            with patch("moban.plugins.BaseEngine.render_to_files") as fake:
+                main()
+                call_args = list(fake.call_args[0][0])
+                eq_(
+                    call_args,
+                    [
+                        ("README.rst.jj2", "custom-data.yaml", "README.rst"),
+                        ("setup.py.jj2", "data.yml", "setup.py"),
+                    ],
+                )
 
     def tearDown(self):
         os.unlink(self.config_file)
         os.unlink(self.data_file)
-        self.patcher.stop()
+        self.patcher1.stop()
 
 
 class TestTemplateTypeOption:
@@ -332,10 +366,12 @@ class TestTemplateTypeOption:
         with open(self.config_file, "w") as f:
             f.write("hello: world")
 
-    @patch("moban.engine.Engine.render_to_file")
+    @patch("moban.plugins.BaseEngine.render_to_file")
     def test_mako_option(self, fake_template_doer):
         test_args = ["moban", "-t", "a.mako"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             fake_template_doer.assert_called_with(
                 "a.mako", "data.yml", "moban.output"