+++ tests/test_utils.py
@@ -3,12 +3,18 @@ import stat
 from shutil import rmtree
 
 from mock import patch
-from nose.tools import eq_
-
-from moban.utils import file_permissions_copy
-from moban.utils import write_file_out
-from moban.utils import strip_off_trailing_new_lines
-from moban.utils import mkdir_p, expand_directories
+from nose.tools import eq_, raises
+from moban.utils import (
+    mkdir_p,
+    get_repo_name,
+    write_file_out,
+    file_permissions,
+    get_template_path,
+    expand_directories,
+    file_permissions_copy,
+    strip_off_trailing_new_lines,
+)
+from moban.exceptions import FileNotFound
 
 
 def create_file(test_file, permission):
@@ -30,6 +36,11 @@ def test_file_permission_copy():
     )
     os.unlink(test_source)
     os.unlink(test_dest)
+
+
+@raises(FileNotFound)
+def test_file_permissions_file_not_found():
+    file_permissions("I does not exist")
 
 
 def test_file_permission_copy_symlink():
@@ -85,6 +96,14 @@ def test_expand_dir():
     eq_(results, expected)
 
 
+def test_get_template_path():
+    temp_dirs = ["tests/fixtures/template-tests", "tests/abc", "tests/abc"]
+    template = "a.jj2"
+    template_path = get_template_path(temp_dirs, template)
+    expected = os.path.join(os.getcwd(), "tests/fixtures/template-tests/a.jj2")
+    eq_(template_path, expected)
+
+
 @patch("subprocess.check_call")
 def test_pip_install(fake_check_all):
     import sys
@@ -94,3 +113,45 @@ def test_pip_install(fake_check_all):
     fake_check_all.assert_called_with(
         [sys.executable, "-m", "pip", "install", "package1 package2"]
     )
+
+
+@patch("subprocess.check_call")
+def test_git_clone(fake_check_all):
+    from moban.utils import git_clone
+
+    git_clone(["https://github.com/my/repo", "https://gitlab.com/my/repo"])
+    fake_check_all.assert_called_with(
+        ["git", "clone", "https://gitlab.com/my/repo", "repo"]
+    )
+
+
+@patch("os.chdir")
+@patch("subprocess.check_call")
+def test_git_clone_with_submodules(fake_check_all, _):
+    from moban.utils import git_clone
+
+    git_clone(
+        ["https://github.com/my/repo", "https://gitlab.com/my/repo"],
+        submodule=True,
+    )
+    fake_check_all.assert_called_with(["git", "submodule", "update"])
+
+
+@patch("os.path.exists", return_value=True)
+@patch("os.chdir")
+@patch("subprocess.check_call")
+def test_git_clone_with_existing_repo(fake_check_all, _, __):
+    from moban.utils import git_clone
+
+    git_clone(
+        ["https://github.com/my/repo", "https://gitlab.com/my/repo"],
+        submodule=True,
+    )
+    fake_check_all.assert_called_with(["git", "submodule", "update"])
+
+
+def test_get_repo_name():
+    repos = ["https://github.com/abc/repo", "https://github.com/abc/repo/"]
+    actual = [get_repo_name(repo) for repo in repos]
+    expected = ["repo", "repo"]
+    eq_(expected, actual)